+++ projects/arlas-components/src/lib/components/results/result-list/result-list.component.ts
@@ -40,6 +40,7 @@ import { ThumbnailFitEnum } from '../utils/enumerations/thumbnailFitEnum';
 import { Action, ElementIdentifier, FieldsConfiguration, ItemDataType,
   PageQuery, ResultListOptions, matchAndReplace } from '../utils/results.utils';
 import { marker } from '@biesbjerg/ngx-translate-extract-marker';
+import { ResultlistNotifierService } from '../../../services/resultlist.notifier.service';
 
 /**
  * ResultList component allows to structure data in a filterable and sortable table.
@@ -221,10 +222,16 @@ export class ResultListComponent implements OnInit, DoCheck, OnChanges, AfterVie
   @Input() public indeterminatedItems: Set<string> = new Set<string>();
 
   /**
- * @Input : Angular
- * @description List of items ids that are in a selected status.
- */
+   * @Input : Angular
+   * @description List of items ids that are in a selected status.
+  */
   @Input() public selectedItems: Set<string> = new Set<string>();
+
+  /**
+   * @Input : Angular
+   * @description Map <itemId, Set<actionIds>> : for each item, gives the list of activated actions.
+  */
+  @Input() public activatedActionsPerItem: Map<string, Set<string>> = new Map<string, Set<string>>();
 
   /**
    * @Input : Angular
@@ -447,7 +454,13 @@ export class ResultListComponent implements OnInit, DoCheck, OnChanges, AfterVie
    * @Output : Angular
    * @description Emits when result list is updated.
    */
-  @Output()  public onResultListUpdate: EventEmitter<boolean> = new EventEmitter<boolean>();
+  @Output() public onResultListUpdate: EventEmitter<boolean> = new EventEmitter<boolean>();
+
+  /**
+   * @Output : Angular
+   * @description Emits when the list is ready.
+   */
+  @Output() public onListLoaded = new EventEmitter<boolean>();
 
   public columns: Array<Column>;
   public items: Array<Item> = new Array<Item>();
@@ -483,7 +496,7 @@ export class ResultListComponent implements OnInit, DoCheck, OnChanges, AfterVie
 
 
   public constructor(iterableRowsDiffer: IterableDiffers, iterableColumnsDiffer: IterableDiffers, private el: ElementRef,
-    private colorService: ArlasColorService, public translate: TranslateService,
+    private colorService: ArlasColorService, public translate: TranslateService, private notifier: ResultlistNotifierService,
     private cdr: ChangeDetectorRef) {
     this.iterableRowsDiffer = iterableRowsDiffer.find([]).create(null);
     this.iterableColumnsDiffer = iterableColumnsDiffer.find([]).create(null);
@@ -521,6 +534,7 @@ export class ResultListComponent implements OnInit, DoCheck, OnChanges, AfterVie
   }
 
   public ngAfterViewInit(): void {
+    this.onListLoaded.next(true);
     this.setTableWidth();
     this.setTableHeight();
   }
@@ -547,8 +561,16 @@ export class ResultListComponent implements OnInit, DoCheck, OnChanges, AfterVie
       // Reset selected items when data change (ie a filter is applied/removed or pagination occur)
       this.selectedItems = new Set<string>();
       this.isPreviousPageRequested = false;
+
+      // If the selected item is not in the current list of items, close the detail
+      const selectedItemInData = !!this.selectedGridItem && this.rowItemList
+        .map(item => <string>item.get(this.fieldsConfiguration.idFieldName))
+        .includes(this.selectedGridItem.identifier);
+      if (!(!!changes['rowItemList'].currentValue && selectedItemInData)) {
+        this.closeDetail(true);
+      }
+
       this.onChangeItems.next(changes['rowItemList'].currentValue);
-      this.closeDetail(true);
     }
     if (changes['isDetailledGridOpen'] !== undefined) {
       this.isDetailledGridOpen = changes['isDetailledGridOpen'].currentValue;
@@ -785,8 +807,8 @@ export class ResultListComponent implements OnInit, DoCheck, OnChanges, AfterVie
    * @param item hovered item
    */
   public onEnterItem(item: Item): void {
-    this.setItemActions(item);
     this.setConsultedItem(item.identifier);
+    this.notifier.notifyItemHover(item);
   }
 
   /**
@@ -913,26 +935,6 @@ export class ResultListComponent implements OnInit, DoCheck, OnChanges, AfterVie
     });
   }
 
-  /**
-   * @description set the list of actions of an item
-   * @param item
-   */
-  public setItemActions(item: Item): void {
-    if (item && (!item.actions || (item.actions && item.actions.length === 0))) {
-      item.actions = new Array<Action>();
-      this.detailedDataRetriever.getActions(item).subscribe(actions => {
-        actions.forEach(action => {
-          item.actions.push({
-            id: action.id,
-            label: action.label,
-            actionBus: action.actionBus,
-            cssClass: action.cssClass,
-            tooltip: action.tooltip
-          });
-        });
-      });
-    }
-  }
 
   public byFieldName(item1: Column, item2: Column) {
     return item1 && item2 ? item1.fieldName === item2.fieldName : item1 === item2;