+++ projects/arlas-map/src/lib/map-settings/map-settings.component.ts
@@ -83,8 +83,8 @@ export class MapSettingsDialogComponent implements OnInit {
    * - the operation ("within", "intersects", "notwithin", "notintersects")
    * - the geometry field to query
    */
-  @Output() public geoQueryEmitter = new Subject<Map<string, GeoQuery>>();
-  public emittedGeoQueries = new Map<string, GeoQuery>();
+  @Output() public geoQueryEmitter: Subject<Map<string, GeoQuery>> = new Subject<Map<string, GeoQuery>>();
+  public emittedGeoQueries: Map<string, GeoQuery> = new Map();
   /** Constants */
   public GEO_QUERIES_DESCRIPTION = marker('Draw a bbox or a polygon that');