+++ CHANGELOG.md
@@ -1,5 +1,57 @@
 # Change Log
 
+## [v27.0.0-rc.1](https://github.com/gisaia/ARLAS-web-components/tree/v27.0.0-rc.1) (2024-12-03)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.12...v27.0.0-rc.1)
+
+## [v26.0.12](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.12) (2024-12-02)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.11...v26.0.12)
+
+## [v26.0.11](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.11) (2024-11-26)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v27.0.0-beta.2...v26.0.11)
+
+**Fixed bugs:**
+
+- actions update needs existence verification [\#888](https://github.com/gisaia/ARLAS-web-components/issues/888)
+
+## [v27.0.0-beta.2](https://github.com/gisaia/ARLAS-web-components/tree/v27.0.0-beta.2) (2024-11-14)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.10...v27.0.0-beta.2)
+
+## [v26.0.10](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.10) (2024-11-12)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.9...v26.0.10)
+
+## [v26.0.9](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.9) (2024-11-12)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.8...v26.0.9)
+
+## [v26.0.8](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.8) (2024-11-12)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.7...v26.0.8)
+
+**Fixed bugs:**
+
+- Legend can sometimes not display fully [\#881](https://github.com/gisaia/ARLAS-web-components/issues/881) [[MAP](https://github.com/gisaia/ARLAS-web-components/labels/MAP)]
+
+## [v26.0.7](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.7) (2024-11-06)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.6...v26.0.7)
+
+## [v26.0.6](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.6) (2024-10-30)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v27.0.0-beta.1...v26.0.6)
+
+## [v27.0.0-beta.1](https://github.com/gisaia/ARLAS-web-components/tree/v27.0.0-beta.1) (2024-10-29)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.6-beta.1...v27.0.0-beta.1)
+
+## [v26.0.6-beta.1](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.6-beta.1) (2024-09-30)
+
+[Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.5...v26.0.6-beta.1)
+
 ## [v26.0.5](https://github.com/gisaia/ARLAS-web-components/tree/v26.0.5) (2024-09-22)
 
 [Full Changelog](https://github.com/gisaia/ARLAS-web-components/compare/v26.0.4...v26.0.5)